Rename and document test.bytecode_helper as test.support.bytecode_helper
